+++ test/cli.js
@@ -1,4 +1,4 @@
-var bag = require('bagofholding'),
+var bag = require('bagofcli'),
   buster = require('buster'),
   cli = require('../lib/cli'),
   Couchpenter = require('../lib/couchpenter');
@@ -24,21 +24,18 @@ buster.testCase('cli - exec', {
       assert.defined(actions.commands['live-deploy-view'].action);
       done();
     };
-    this.stub(bag, 'cli', { command: mockCommand });
+    this.stub(bag, 'command', mockCommand);
     cli.exec();
   }
 });
 
 buster.testCase('cli - init', {
   'should contain init command and delegate to couchpenter init when exec is called': function (done) {
-    this.stub(bag, 'cli', {
-      command: function (base, actions) {
-        actions.commands.init.action();
-      },
-      exit: bag.cli.exit
+    this.stub(bag, 'command', function (base, actions) {
+      actions.commands.init.action();
     });
     this.stub(Couchpenter.prototype, 'init', function (cb) {
-      assert.equals(typeof bag.cli.exit, 'function');
+      assert.equals(typeof bag.exit, 'function');
       done();
     });
     cli.exec();
@@ -58,12 +55,11 @@ buster.testCase('cli - task', {
       self.stub(Couchpenter.prototype, couchpenterFn, function (cb) {
         cb(null, [{ id: 'id1', message: 'someresult1' }, { id: 'id2', message: 'someresult2', error: { status_code: 404 } }]);
       });
-      self.stub(bag, 'cli', {
-        command: function (base, actions) {
-          actions.commands[command].action({ url: 'http://someurl', setupFile: 'somesetupfile', dir: 'somedir', interval: 2000 });
-        },
-        exit: bag.cli.exit,
-        exitCb: bag.cli.exitCb
+      if (bag.command.restore) {
+        bag.command.restore();
+      }
+      self.stub(bag, 'command', function (base, actions) {
+        actions.commands[command].action({ url: 'http://someurl', setupFile: 'somesetupfile', dir: 'somedir', interval: 2000 });
       });
       cli.exec();
     };