+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.4-pre
-*
+* Ignore proxy when CouchDB URL hostname is either localhost or 127.0.0.1
+* Replace bagofholding with bagofcli and bagofrequest
 
 ### 0.2.3
 * Add live-deploy-view target couchpenter#liveDeployView (Gary Chin)